// src/domain/enums/education-level.enum.ts
export enum EducationLevel {
	Elementary = "ELEMENTARY",
	HighSchool = "HIGH_SCHOOL",
	Technical = "TECHNICAL",
	Technologist = "TECHNOLOGIST",
	Bachelor = "BACHELOR",
	PostGraduate = "POST_GRADUATE",
	Master = "MASTER",
	Doctorate = "DOCTORATE",
	PostDoctorate = "POST_DOCTORATE",

	Unknown = "UNKNOWN",
}