// src/platform/gupy/types/education/enum/gupy.education.enum.ts
export enum GupyEducationTypes {
	technical_course = "technical_course",
	technological = "technological",
	graduation = "graduation",
	post_graduate = "post_graduate",
	master_degree = "master_Degree",
	phd = "phd",
}

export enum GupyEducationConclusionStatus {
	education_complete = "complete",
	education_in_progress = "in_progress",
	education_incomplete = "incomplete",
}

export enum GupyUnderGraduationTypes {
	completedHighSchool = "completed_high_school",
	inProgressHighSchool = "in_progress_high_school",
	incompleteHighSchool = "incompleted_high_school",
	completedElementarySchool = "completed_elementary_school",
	inProgressFundamental = "in_progress_elementary_school",
	incompleteElementarySchool = "incompleted_elementary_school",
}